[OM-59798]: Add retry logic to DbEndpoint initialization
Review RB-39870: https://rbcommons.com/s/VMTurbo/r/39870

DbEndpoint initialization may fail on first attempt due to a variety
of self-correcting problems, like a the target DB is running in a pod
that is still starting up. This change tolerates such initial failures
by retrying initialization on first attempted use of the endpoint,
until initialization succeeds or the configured retry schedule has
been exhausted.

+++ com.vmturbo.sql.test.utils/src/main/java/com/vmturbo/sql/utils/DbEndpointTestRule.java
@@ -4,12 +4,12 @@ import static com.vmturbo.sql.utils.DbEndpointResolver.COMPONENT_TYPE_PROPERTY;
 import static com.vmturbo.sql.utils.DbEndpointResolver.DB_HOST_PROPERTY;
 import static com.vmturbo.sql.utils.DbEndpointResolver.DB_NAME_SUFFIX_PROPERTY;
 import static com.vmturbo.sql.utils.DbEndpointResolver.DB_PORT_PROPERTY;
+import static com.vmturbo.sql.utils.DbEndpointResolver.DB_RETRY_BACKOFF_TIMES_SEC_PROPERTY;
 import static com.vmturbo.sql.utils.DbEndpointResolver.taggedPropertyName;
 import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import java.sql.SQLException;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
@@ -98,6 +98,9 @@ public class DbEndpointTestRule implements TestRule {
             String componentName, Map<String, String> propertySettings, DbEndpoint... endpoints) {
         this.propertySettings = new HashMap<>(propertySettings);
         this.propertySettings.put(COMPONENT_TYPE_PROPERTY, componentName);
+        // put in a safe retry schedule for tests (the normal default will wait for ever)
+        // this retries every 3 seconds for up to 30 seconds
+        this.propertySettings.put(DB_RETRY_BACKOFF_TIMES_SEC_PROPERTY, "0,3,3,3,3,3,3,3,3,3,3");
         this.endpoints = Arrays.asList(endpoints);
         final String provisioningSuffix = "_" + System.currentTimeMillis();
         for (DbEndpoint endpoint : this.endpoints) {
@@ -130,12 +133,17 @@ public class DbEndpointTestRule implements TestRule {
     }
 
     protected void before(final Description description) throws
-            UnsupportedDialectException, SQLException {
+            Throwable {
         logger.info("Setting up temporary DB and/or truncating data for test " + description.getMethodName());
         DbEndpointCompleter.setResolver(propertySettings::get, mockPasswordUtil(), false);
         for (DbEndpoint endpoint : endpoints) {
             if (!endpoint.isReady()) {
-                DbEndpointCompleter.completePendingEndpoint(endpoint);
+                try {
+                    DbEndpointCompleter.completePendingEndpoint(endpoint);
+                } catch (Exception e) {
+                    logger.warn("Endpoint {} initialization failed; entering retry loop", endpoint, e);
+                }
+                endpoint.awaitCompletion();
             }
             if (endpoint.getConfig().getDbAccess().isWriteAccess()) {
                 endpoint.getAdapter().truncateAllTables();
@@ -143,11 +151,13 @@ public class DbEndpointTestRule implements TestRule {
         }
     }
 
-    private void afterClass() {
+    private void afterClass() throws Throwable {
         logger.info("Finished tests, dropping temporary databases");
         for (final DbEndpoint endpoint : endpoints) {
-            endpoint.getAdapter().dropDatabase();
-            endpoint.getAdapter().dropUser();
+            if (endpoint.isReady()) {
+                endpoint.getAdapter().dropDatabase();
+                endpoint.getAdapter().dropUser();
+            }
         }
         // don't allow this rule's configurations to leak into other test classes involving
         // the same endpoints
